+++ src/RecyclableCalculator.Core/Dto/RecyclableItemDtos/RecyclableItemUpdateRequest.cs
@@ -24,12 +24,12 @@ namespace RecyclableCalculator.Core.Dto.RecyclableItemDtos
 		[Required]
 		[DisplayFormat(DataFormatString = "{0:F2}", ApplyFormatInEditMode = true)]
 		[Range(0, int.MaxValue)]
-		public double Weight { get; set; }
+		public decimal Weight { get; set; }
 
 		[Required]
 		[DisplayFormat(DataFormatString = "{0:F2}", ApplyFormatInEditMode = true)]
 		[Range(0, int.MaxValue)]
 		[DisplayName("Computed Rate")]
-		public double ComputedRate { get; set; }
+		public decimal ComputedRate { get; set; }
 	}
 }
